+++ plugin/src/test/groovy/org/openbakery/InfoPlistModifyTaskTest.groovy
@@ -1,8 +1,10 @@
 package org.openbakery
 
+import org.apache.commons.io.FileUtils
 import org.gmock.GMockController
 import org.gradle.api.Project
 import org.gradle.testfixtures.ProjectBuilder
+import org.testng.annotations.AfterMethod
 import org.testng.annotations.BeforeMethod
 import org.testng.annotations.Test
 
@@ -14,6 +16,7 @@ class InfoPlistModifyTaskTest {
 
 	Project project
 	InfoPlistModifyTask task
+	File infoPlist
 
 	GMockController mockControl
 	CommandRunner commandRunnerMock
@@ -23,18 +26,29 @@ class InfoPlistModifyTaskTest {
 		mockControl = new GMockController()
 		commandRunnerMock = mockControl.mock(CommandRunner)
 
-		project = ProjectBuilder.builder().build()
-		project.buildDir = new File('build').absoluteFile
+		File projectDir = new File(System.getProperty("java.io.tmpdir"), "gradle-xcodebuild")
+		project = ProjectBuilder.builder().withProjectDir(projectDir).build()
+		project.buildDir = new File(projectDir, 'build').absoluteFile
+
 		project.apply plugin: org.openbakery.XcodePlugin
-
-		project.xcodebuild.infoPlist = "Info.plist"
-
 
 		task = project.tasks.findByName('infoplist-modify')
 		task.setProperty("commandRunner", commandRunnerMock)
 
 
+		def archiveDirectory = new File(project.getBuildDir(), XcodeBuildArchiveTask.ARCHIVE_FOLDER + "/Example.xcarchive/Products/Applications/Example.app")
 
+		archiveDirectory.mkdirs()
+
+		infoPlist = new File(archiveDirectory, "Info.plist")
+
+		FileUtils.write(infoPlist, "dummy")
+
+	}
+
+	@AfterMethod
+	void cleanUp() {
+		FileUtils.deleteDirectory(project.projectDir)
 	}
 
 
@@ -44,7 +58,7 @@ class InfoPlistModifyTaskTest {
 
 		List<String> commandList
 		commandList?.clear()
-		commandList = ["/usr/libexec/PlistBuddy", "Info.plist", "-c", "Set :CFBundleIdentifier " + project.infoplist.bundleIdentifier]
+		commandList = ["/usr/libexec/PlistBuddy", infoPlist.absolutePath, "-c", "Set :CFBundleIdentifier " + project.infoplist.bundleIdentifier]
 		commandRunnerMock.run(commandList).times(1)
 
 
@@ -58,7 +72,7 @@ class InfoPlistModifyTaskTest {
 	void testModifyCommand_single() {
 		project.infoplist.commands = "Add CFBundleURLTypes:0:CFBundleURLName string"
 
-		def commandList = ["/usr/libexec/PlistBuddy", "Info.plist", "-c", project.infoplist.commands[0]]
+		def commandList = ["/usr/libexec/PlistBuddy", infoPlist.absolutePath, "-c", project.infoplist.commands[0]]
 		commandRunnerMock.run(commandList).times(1)
 
 
@@ -73,11 +87,11 @@ class InfoPlistModifyTaskTest {
 		project.infoplist.commands = ["Add CFBundleURLTypes:0:CFBundleURLName string", "Add CFBundleURLTypes:0:CFBundleURLSchemes array" ]
 
 
-		def commandList = ["/usr/libexec/PlistBuddy", "Info.plist", "-c", project.infoplist.commands[0]]
+		def commandList = ["/usr/libexec/PlistBuddy", infoPlist.absolutePath, "-c", project.infoplist.commands[0]]
 		commandRunnerMock.run(commandList).times(1)
 
 
-		commandList = ["/usr/libexec/PlistBuddy", "Info.plist", "-c", project.infoplist.commands[1]]
+		commandList = ["/usr/libexec/PlistBuddy", infoPlist.absolutePath, "-c", project.infoplist.commands[1]]
 		commandRunnerMock.run(commandList).times(1)
 
 
@@ -94,7 +108,7 @@ class InfoPlistModifyTaskTest {
 
 		List<String> commandList
 		commandList?.clear()
-		commandList = ["/usr/libexec/PlistBuddy", "Info.plist", "-c", "Set :CFBundleVersion 1.0.0"]
+		commandList = ["/usr/libexec/PlistBuddy", infoPlist.absolutePath, "-c", "Set :CFBundleVersion 1.0.0"]
 		commandRunnerMock.run(commandList).times(1)