+++ plugin/src/test/groovy/org/openbakery/XcodeBuildTaskTest.groovy
@@ -364,4 +364,15 @@ class XcodeBuildTaskTest {
 		}
 	}
 
+	@Test
+	void dependsOn() {
+		def dependsOn  = xcodeBuildTask.getDependsOn()
+
+		assert dependsOn.contains(XcodePlugin.XCODE_CONFIG_TASK_NAME)
+		assert dependsOn.contains(XcodePlugin.KEYCHAIN_CREATE_TASK_NAME)
+		assert dependsOn.contains(XcodePlugin.PROVISIONING_INSTALL_TASK_NAME)
+
+		assert !dependsOn.contains(XcodePlugin.INFOPLIST_MODIFY_TASK_NAME)
+	}
+
 }